Cleaned up mostly and moved old testing code to separte file
